+++ apps/framework-cli/src/cli/routines.rs
@@ -119,7 +119,6 @@ use crate::framework::core::plan::plan_changes;
 use crate::framework::core::plan::InfraPlan;
 use crate::framework::core::primitive_map::PrimitiveMap;
 use crate::framework::core::state_storage::StateStorageBuilder;
-use crate::infrastructure::olap::clickhouse::config::parse_clickhouse_connection_string;
 use crate::infrastructure::olap::clickhouse::{check_ready, create_client};
 use crate::infrastructure::olap::OlapOperations;
 use crate::infrastructure::orchestration::temporal_client::{
@@ -1097,18 +1096,10 @@ pub enum RemoteSource<'a> {
 }
 
 pub async fn remote_gen_migration(
-    project: &mut Project,
+    project: &Project,
     remote: RemoteSource<'_>,
 ) -> anyhow::Result<MigrationPlanWithBeforeAfter> {
     use anyhow::Context;
-
-    // For serverless migrations, override project config BEFORE loading local inframap
-    // This ensures the local inframap is built with the correct database from the CLI URL
-    if let RemoteSource::Serverless { clickhouse_url, .. } = remote {
-        project.clickhouse_config = parse_clickhouse_connection_string(clickhouse_url)
-            .context("Failed to parse ClickHouse URL")?;
-        info!("Aligned project clickhouse config with remote clickhouse url");
-    }
 
     // Build the inframap from the local project
     let local_infra_map = if project.features.data_model_v2 {
